Club main and event_details
