+++ event-manager/README.md
@@ -23,13 +23,13 @@ This script creates events across -
 
    3.2 Guild ID [`DISCORD_GUILD_ID`] (developersIndia => `1229786646468362260`)
 
-   3.3 Calender Id [`GOOGLE_CALENDER_ID`] (developerIndia => `9f1337e4154910eb1bdb3bfac32b88f69546468b1281a6db58f50a909df5049f@group.calendar.google.com`)
+   3.3 calendar Id [`GOOGLE_calendar_ID`] (developerIndia => `9f1337e4154910eb1bdb3bfac32b88f69546468b1281a6db58f50a909df5049f@group.calendar.google.com`)
 
-4. Connect Google Calender through [Google cloud Console](https://console.cloud.google.com/)
+4. Connect Google calendar through [Google cloud Console](https://console.cloud.google.com/)
 
     4.1 Create a Project on Google Cloud Console
    
-    4.2 Search for Calender API and enable it
+    4.2 Search for calendar API and enable it
 
     4.3 Create Credentials ->  OAuth Client ID -> Application type as Desktop
