+++ event-manager/main.py
@@ -1,12 +1,31 @@
-# SCRIPT
+import os
+from dotenv import load_dotenv
 
-# Execute event_details.py
-import event_details
+load_dotenv()
+bot_token = os.getenv('DISCORD_BOT_TOKEN')
+guild_id = int(os.getenv('DISCORD_GUILD_ID'))
+calendar_id = os.getenv('GOOGLE_CALENDAR_ID')
 
-# Execute discord_.py
-import discord_
+def get_input():
+    NAME = input("Event title: ")
+    DESCRIPTION = input("Event description: ")
+    LOCATION = input("Event location (red -> for developersindia subreddit): ")
+    if LOCATION == "red":
+        LOCATION = "https://www.reddit.com/r/developersindia/"
+    DATE = input("Enter the event date (yyyy-mm-dd): ")
+    START_TIME = input("Enter the event start time (hh:mm)- ")
+    END_TIME = input("Enter the event end time (hh:mm)- ")
 
-# Execute google_calender.py
-import google_calender_
+    return NAME, DESCRIPTION, LOCATION, DATE, START_TIME, END_TIME
+
+EVENT_NAME, EVENT_DESCRIPTION, EVENT_LOCATION, EVENT_DATE, EVENT_START_TIME, EVENT_END_TIME = get_input()
+
+# Execute create_discord_event
+from discord_ import create_discord_event
+create_discord_event(bot_token, guild_id, EVENT_NAME, EVENT_DESCRIPTION, EVENT_LOCATION, EVENT_DATE, EVENT_START_TIME, EVENT_END_TIME)
+
+# Execute google_calendar_event
+from google_calendar_ import create_google_calendar_event
+create_google_calendar_event(calendar_id, EVENT_NAME, EVENT_DESCRIPTION, EVENT_LOCATION, EVENT_DATE, EVENT_START_TIME, EVENT_END_TIME)
 
 # END
